Fixed bugs and warnings in Buffer, Reader, Writer and Database

+++ Cereal/Cereal/src/Database.h
@@ -1,4 +1,4 @@
-//  Cereal: A C++ Serialization library
+//  Cereal: A C++/C# Serialization library
 //  Copyright (C) 2016  The Cereal Team
 //
 //  This program is free software: you can redistribute it and/or modify
@@ -107,9 +107,9 @@ namespace Cereal {
 			return true;
 		}
 
-		inline unsigned long long getSize() const
+		inline unsigned int getSize() const
 		{
-			unsigned long long ret = sizeof(short);
+			unsigned int ret = sizeof(short);
 
 			switch (version)
 			{
